♻️ Migrate the code into the new sub-package *gearmeshing_ai.info_provider.mcp*.
