+++ gearmeshing_ai/info_provider/mcp/gateway_api/models/domain.py
@@ -12,8 +12,8 @@ from typing import Any, Dict, List, Optional
 
 from pydantic import AnyHttpUrl, Field
 
-from gearmeshing_ai.info_provider.schemas.base import BaseSchema
-from gearmeshing_ai.info_provider.schemas.core import (
+from ...schemas.base import BaseSchema
+from ...schemas.core import (
     McpServerRef,
     ServerKind,
     TransportType,
