+++ gearmeshing_ai/info_provider/mcp/strategy/models/dto.py
@@ -19,8 +19,8 @@ from typing import Any, Callable, Dict, List, Optional, cast
 
 from pydantic import ConfigDict, Field, model_validator
 
-from gearmeshing_ai.info_provider.schemas.base import BaseSchema
-from gearmeshing_ai.info_provider.schemas.core import McpTool, ToolArgument, ToolCallResult
+from ...schemas.base import BaseSchema
+from ...schemas.core import McpTool, ToolArgument, ToolCallResult
 
 # Use a descriptive alias without recursive typing to avoid Pydantic recursion issues
 JSONValue = Any
