+++ test/contract_test/info_provider/test_client_protocols_contract.py
@@ -5,14 +5,14 @@ from typing import Any, AsyncIterator, Dict, List, Optional
 
 import pytest
 
-from gearmeshing_ai.info_provider.base import (
+from gearmeshing_ai.info_provider.mcp.base import (
     AsyncMCPInfoProvider,
     ClientCommonMixin,
     MCPInfoProvider,
 )
-from gearmeshing_ai.info_provider.policy import ToolPolicy
-from gearmeshing_ai.info_provider.schemas.config import McpClientConfig
-from gearmeshing_ai.info_provider.schemas.core import (
+from gearmeshing_ai.info_provider.mcp.policy import ToolPolicy
+from gearmeshing_ai.info_provider.mcp.schemas.config import McpClientConfig
+from gearmeshing_ai.info_provider.mcp.schemas.core import (
     McpServerRef,
     McpTool,
     ServerKind,
