+++ test/integration_test/info_provider/test_mcp_client_integration.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, List
 
 import httpx
 
-from gearmeshing_ai.info_provider.client_sync import McpClient
-from gearmeshing_ai.info_provider.schemas.config import (
+from gearmeshing_ai.info_provider.mcp.client_sync import McpClient
+from gearmeshing_ai.info_provider.mcp.schemas.config import (
     GatewayConfig,
     McpClientConfig,
     ServerConfig,
