+++ test/unit_test/info_provider/strategy/test_gateway_async_strategy.py
@@ -6,8 +6,8 @@ from typing import Any, Dict, List
 import httpx
 import pytest
 
-from gearmeshing_ai.info_provider.gateway_api.client import GatewayApiClient
-from gearmeshing_ai.info_provider.strategy.gateway_async import AsyncGatewayMcpStrategy
+from gearmeshing_ai.info_provider.mcp.gateway_api.client import GatewayApiClient
+from gearmeshing_ai.info_provider.mcp.strategy.gateway_async import AsyncGatewayMcpStrategy
 
 
 def _mock_transport(state: dict) -> httpx.MockTransport:
