+++ test/unit_test/info_provider/strategy/test_gateway_strategy_cache_and_auth.py
@@ -5,8 +5,8 @@ from typing import Any, Dict, List
 
 import httpx
 
-from gearmeshing_ai.info_provider.gateway_api.client import GatewayApiClient
-from gearmeshing_ai.info_provider.strategy.gateway import GatewayMcpStrategy
+from gearmeshing_ai.info_provider.mcp.gateway_api.client import GatewayApiClient
+from gearmeshing_ai.info_provider.mcp.strategy.gateway import GatewayMcpStrategy
 
 
 def _mock_transport(state: dict) -> httpx.MockTransport:
