+++ test/unit_test/info_provider/strategy/test_pagination_direct_async_strategy.py
@@ -5,8 +5,8 @@ from typing import Any, Dict, List
 import httpx
 import pytest
 
-from gearmeshing_ai.info_provider.schemas.config import ServerConfig
-from gearmeshing_ai.info_provider.strategy.direct_async import AsyncDirectMcpStrategy
+from gearmeshing_ai.info_provider.mcp.schemas.config import ServerConfig
+from gearmeshing_ai.info_provider.mcp.strategy.direct_async import AsyncDirectMcpStrategy
 
 
 def _mock_transport(state: dict) -> httpx.MockTransport:
