+++ test/unit_test/info_provider/strategy/test_pagination_gateway_strategy.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 import httpx
 
-from gearmeshing_ai.info_provider.gateway_api.client import GatewayApiClient
-from gearmeshing_ai.info_provider.strategy.gateway import GatewayMcpStrategy
+from gearmeshing_ai.info_provider.mcp.gateway_api.client import GatewayApiClient
+from gearmeshing_ai.info_provider.mcp.strategy.gateway import GatewayMcpStrategy
 
 
 def _mock_transport_paginated() -> httpx.MockTransport:
