+++ test/unit_test/info_provider/test_policy_read_only.py
@@ -5,10 +5,10 @@ from typing import Any, Dict, List
 import httpx
 import pytest
 
-from gearmeshing_ai.info_provider.client_sync import McpClient
-from gearmeshing_ai.info_provider.errors import ToolAccessDeniedError
-from gearmeshing_ai.info_provider.policy import ToolPolicy
-from gearmeshing_ai.info_provider.schemas.config import McpClientConfig, ServerConfig
+from gearmeshing_ai.info_provider.mcp.client_sync import McpClient
+from gearmeshing_ai.info_provider.mcp.errors import ToolAccessDeniedError
+from gearmeshing_ai.info_provider.mcp.policy import ToolPolicy
+from gearmeshing_ai.info_provider.mcp.schemas.config import McpClientConfig, ServerConfig
 
 
 def _mock_transport() -> httpx.MockTransport:
